Added in-app generation of a deck and very basic functionality all in a separate VC than the parse-based workout.
I set up a Scene heirarchy so the app loads on to a Deck selection
scene with 2 decks. One is the Parse-based deck (DeckTest), and the
other is the in-app deck generator (OfflineTest). Both of these decks
are just UIButton objects for now.
I added placeholder scenes for Help and Settings. These can navigated
to and from, but have no content or function yet.
I used the existing model classes and just commented my begins and
ends. This should show us where the logical intersections are for the
app with what is already done on Parse.
The Workout scene includes an (unsanitized) timer, left and right swipe
gestures on the placehlder card image to advance the workout, and a
progress bar (and label) that show user advancement through the deck.

This is a very rough pass just to get the basics working.